+++ src/main/java/ru/otus/spring/service/AuthorServiceImpl.java
@@ -43,7 +43,7 @@ public class AuthorServiceImpl implements AuthorService, InitializingBean {
 	}
 
 	@Override
-	public void deleteById(int id) {
+	public void deleteById(Long id) {
 		authorDao.deleteById(id);
 		authors = authorDao.getAll().stream().collect(Collectors.toConcurrentMap(Author::getName, a->a));
 	}
